Moved sidebar related components to a single directory
